+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 CHANGELOG
 ---------
 - **Unreleased**
+
+- **December.15.24**: 4.2.0
   - [PR #359](https://github.com/caxlsx/caxlsx/pull/359) Add `PivotTable#grand_totals` option to remove grand totals row/col
   - [PR #362](https://github.com/caxlsx/caxlsx/pull/362) Use widest width even if provided as fixed value
   - [PR #398](https://github.com/caxlsx/caxlsx/pull/398) Add `Axlsx#uri_parser` method for RFC2396 compatibility